Add files via upload
Combined a lot of building types which is an important feature.
"Third place" is an idea in urban design indicating spots besides work or living. Could be civic buildings, churches, bars, parks etc. But here I use it to indicate anything non-commercial, residential or office.
I'm going to spend more time on neighborhoods with "PLEX" housing because in other cities that's a strong indicator of high neighborhood price.